Extracted mapping to jira/mapping
- Created `jira.Mapper` interface
- Implemented the interface with `jira/mapping.Mapper`
- Updated tests to cover `sync` with a mock mapper
- Created test scaffold for `jira/mapping.Mapper`
